Move Cr2Header implementation to a separate .cpp file. Rename the existing internal header file
